telegram_bot/requirements.txt
python-telegram-bot
httpx
python-dotenv
google-cloud-pubsub 